webkit/LayoutTests/platform/blackberry/fast/events/pass-escape-key-to-textarea-expected.txt

This tests that the escape keyboard key, triggered by the BlackBerry Escape (back) button, is passed to a <textarea>. This test PASSED if you see a PASS message below. Otherwise, it FAILED.


PASS keyCode is 27.

